refactor(scrollbar): rebase
