+++ assets/styles/base.less
@@ -606,6 +606,11 @@ a.link {
 .editable-container {
   .editable-input {
     .chat-row-content {
+      display: flex;
+      flex-direction: column;
+      justify-content: flex-start;
+      align-items: flex-start;
+      box-sizing: border-box;
       .md-url {
         &:extend(.font-flair);
       }
