+++ components/ui/SimpleScroll/SimpleScroll.less
@@ -5,6 +5,9 @@
   .wrapper {
     width: @full;
     height: @full;
+    margin: auto;
+    position: relative;
+    white-space: normal;
     box-sizing: border-box;
   }
