+++ components/views/chat/chatbar/Chatbar.less
@@ -101,9 +101,34 @@
       resize: none;
       font-size: @text-size;
       word-break: break-all;
+
+      &.editable-container {
+        > div {
+          padding: 14px 0;
+        }
+      }
+      blockquote {
+        border-left: 4px solid @text-muted;
+        padding-left: @light-spacing;
+      }
+      p {
+        font-size: @text-size !important;
+        .chatbar-tag {
+          &:extend(.round-corners);
+          background: @midground;
+          padding: @xlight-spacing;
+        }
+      }
     }
 
     .editable-container {
+      white-space: normal;
+      display: flex;
+      flex-direction: row;
+      justify-content: flex-start;
+      align-items: center;
+      padding-top: @light-spacing;
+      padding-bottom: @light-spacing;
       min-height: @chatbar-input-height;
       max-height: 50vh;
     }
