+++ components/views/chat/chatbar/editable/Editable.html
@@ -1,0 +1,26 @@
+<UiSimpleScroll
+  scrollMode="vertical"
+  scrollShow="scroll"
+  containerClass="editable-container"
+>
+  <div v-if="value.length === 0" class="placeholder">{{ placeholder }}</div>
+  <div
+    ref="editable"
+    :contenteditable="enabled"
+    autocapitalize="off"
+    class="editable-input"
+    data-cy="editable-input"
+    @selectstart="onSelectStart"
+    @input="onInput"
+    @keydown="handleInputKeydown"
+    @keyup="handleInputKeyup"
+    @paste="handleInputPaste"
+    @drop="handleDrop"
+    @focus="onFocus"
+    @blur="onBlur"
+  >
+    <div class="chat-row-content">
+      <span><br /></span>
+    </div>
+  </div>
+</UiSimpleScroll>
