+++ components/views/chat/chatbar/editable/Editable.less
@@ -1,0 +1,51 @@
+// non-Retina-specific stuff here
+@media not screen and (-webkit-min-device-pixel-ratio: 2),
+  not screen and (min--moz-device-pixel-ratio: 2),
+  not screen and (-o-min-device-pixel-ratio: 2/1),
+  not screen and (min-device-pixel-ratio: 2),
+  not screen and (min-resolution: 192dpi),
+  not screen and (min-resolution: 2dppx) {
+  .editable-container {
+    .chat-row-content {
+      // Chrome bug for emoji spacing on non-retina screens
+      .emoji {
+        letter-spacing: 4px;
+      }
+    }
+  }
+}
+
+.editable-container {
+  position: relative;
+  width: 100%;
+
+  .placeholder {
+    color: @text-muted;
+    position: absolute;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    left: 0;
+    right: 0;
+    &:extend(.background-layer);
+    &:extend(.no-select);
+  }
+
+  .editable-input {
+    width: 100%;
+
+    .chat-row-content {
+      .emoji {
+        font-style: initial;
+      }
+
+      .md-symbol {
+        color: @gray;
+      }
+
+      .md-lang {
+        color: @green;
+      }
+    }
+  }
+}
