+++ components/views/chat/message/edit/Edit.less
@@ -32,7 +32,7 @@
     }
 
     .editable-container {
-      max-height: 50vh;
+      max-height: 25vh;
     }
   }
   &-bottom {
