+++ components/views/chat/search/result/Result.less
@@ -13,6 +13,10 @@
   border: @light-border !important;
   &:extend(.first-layer);
   font-size: @text-size;
+  height: @full;
+  overflow: hidden;
+  box-sizing: border-box;
+  max-height: calc(100vh - 25rem);
 
   &-header {
     font-size: @title-text;
@@ -78,7 +82,8 @@
     }
   }
   &-content {
-    height: calc(100vh - 32rem) !important;
+    height: @full;
+    max-height: calc(100vh - 32rem);
   }
   &-inner {
     display: flex;
