+++ components/views/files/upload/filePreview/FilePreview.html
@@ -21,7 +21,7 @@
     </div>
   </div>
   <template v-if="files.length > 0">
-    <UiScroll horizontalScroll scrollbarVisibility="scroll">
+    <UiSimpleScroll scrollMode="horizontal" scrollShow="scroll">
       <div class="file-list">
         <FilesUploadPreviewItem
           :item="item"
@@ -31,6 +31,6 @@
           v-for="(item, index) in files"
         ></FilesUploadPreviewItem>
       </div>
-    </UiScroll>
+    </UiSimpleScroll>
   </template>
 </div>
