+++ components/views/navigation/server/sidebar/Sidebar.html
@@ -18,9 +18,9 @@
       />
     </div>
     <div class="scrolling hidden-scroll" v-scroll-lock="true">
-      <UiScroll verticalScroll scrollbarVisibility="scroll" enableWrap>
+      <UiSimpleScroll scrollMode="vertical" scrollShow="scroll">
         <ServerSidebarChannels />
-      </UiScroll>
+      </UiSimpleScroll>
     </div>
   </div>
   <div class="filler"></div>
