+++ components/views/navigation/toolbar/pinned/Pinned.html
@@ -5,7 +5,7 @@
 }"
 >
   <div class="pinned-wrap">
-    <UiScroll verticalScroll scrollbarVisibility="scroll" enableWrap>
+    <UiSimpleScroll scrollMode="vertical" scrollShow="scroll">
       <div class="header">
         <archive-icon size="0.85x" />
         <TypographySubtitle
@@ -31,6 +31,6 @@
         <InteractablesButton size="small" type="dark" text="Jump" />
         <x-icon size="1.2x" />
       </div>
-    </UiScroll>
+    </UiSimpleScroll>
   </div>
 </div>
