+++ layouts/Layout.less
@@ -163,6 +163,7 @@
   }
   .search-result {
     margin-right: calc(@sidebar-size + @servers-size);
+    margin-bottom: @normal-spacing;
   }
 }
