+++ pages/friends/mobile/add/index.vue
@@ -1,6 +1,6 @@
 <template>
   <div id="friends-add" v-scroll-lock="true" class="hidden-scroll">
-    <UiScroll vertical-scroll enable-wrap scrollbar-visibility="scroll">
+    <UiSimpleScroll scroll-mode="vertical" scroll-show="scroll">
       <!--<FriendsNav :setRoute="setRoute" :route="route" />-->
       <div class="columns">
         <div class="column top-bar">
@@ -16,7 +16,7 @@
           <FriendsAdd />
         </div>
       </div>
-    </UiScroll>
+    </UiSimpleScroll>
   </div>
 </template>
